refactor(gax): move checkTelemetryEnabled to util.ts and export it
